Allow running system commands with MP back-end (#953)
