+++ thonny/common.py
@@ -13,6 +13,7 @@ from typing import List, Optional  # @UnusedImport
 import subprocess
 import logging
 import traceback
+from threading import Thread
 
 MESSAGE_MARKER = "\x02"
 
@@ -548,3 +549,45 @@ def get_windows_lnk_target(lnk_file_path):
     result = subprocess.check_output(cmd, universal_newlines=True, timeout=3)
 
     return result.strip()
+
+
+def execute_system_command(cmd):
+    env = dict(os.environ).copy()
+    encoding = "utf-8"
+    env["PYTHONIOENCODING"] = encoding
+    # Make sure this python interpreter and its scripts are available
+    # in PATH
+    update_system_path(env, get_augmented_system_path(get_exe_dirs()))
+    popen_kw = dict(
+        stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True, env=env, universal_newlines=True
+    )
+
+    if sys.version_info >= (3, 6):
+        popen_kw["errors"] = "replace"
+        popen_kw["encoding"] = encoding
+
+    assert cmd.cmd_line.startswith("!")
+    cmd_line = cmd.cmd_line[1:]
+    proc = subprocess.Popen(cmd_line, **popen_kw)
+
+    def copy_stream(source, target):
+        while True:
+            c = source.readline()
+            if c == "":
+                break
+            else:
+                target.write(c)
+                target.flush()
+
+    copy_out = Thread(target=lambda: copy_stream(proc.stdout, sys.stdout), daemon=True)
+    copy_err = Thread(target=lambda: copy_stream(proc.stderr, sys.stderr), daemon=True)
+
+    copy_out.start()
+    copy_err.start()
+    try:
+        proc.wait()
+    except KeyboardInterrupt as e:
+        print(str(e), file=sys.stderr)
+
+    copy_out.join()
+    copy_err.join()
